refactor: clear code
