+++ packages/client/src/store/index.ts
@@ -1,3 +1,4 @@
 export { default as store } from './store'
+export * from './reducers'
+export * from './services'
 export type { RootState, AppStore, AppDispatch } from './store'
-export { setSnackbar } from './reducers'
