+++ packages/client/src/store/services/index.ts
@@ -1,2 +1,2 @@
-export { authAPI } from './authService'
-export { profileApi } from './profileService'
+export * from './authService'
+export * from './profileService'
